Move all clone utilities into a separate file (for clarity)
Only the function that are relevant to the flow of clone and
post-processing remain.

Plain move, no modifications
